api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
redis==5.2.1
rq==1.16.2
supabase==2.15.0
PyJWT[crypto]==2.10.1
pydantic==2.10.6
pydantic-settings==2.8.1
